@@ -1,10 +1,12 @@
 {
   "name": "@ploomes/ploomestsserver",
-  "version": "1.0.3",
+  "version": "1.0.5",
   "description": "",
-  "main": "index.ts",
+  "main": "src/index.ts",
   "scripts": {
-    "build" : "tsc",
+    "build": "tsc",
+    "build:check": "tsc --noEmit",
+    "prepare": "npm run build",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
@@ -12,11 +14,13 @@
   "devDependencies": {
     "@types/express": "^4.16.1",
     "@types/node": "^14.14.22",
+    "ts-node": "^9.1.1",
     "tslint": "^5.12.1",
-    "typescript": "^3.3.3"
+    "typescript": "^3.9.7"
   },
   "dependencies": {
     "@ploomes/ploomeststypes": "^1.0.0",
     "axios": "^0.21.1"
-  }
+  },
+  "files": ["lib"]
 }
